classes separadas em arquivos diferentes. Código está preso já que o interpretador quer chamar da classe TocaDisco uma função da Agulha direto da classe e não pelo objeto instanciado, e obviamente isso dá erro
